requirements.txt
streamlit
langchain-community
langchain-core
langchain-huggingface
langchain-groq
pymupdf
python-dotenv
faiss-cpu
sentence-transformers
